[#9][refactor] 패키지 구조 변경(IntelliJ 이용)
