+++ src/main/java/dev/idion/bladitodo/domain/dto/BoardWithListIdResponse.java
@@ -1,7 +1,7 @@
-package dev.idion.bladitodo.domain.http;
+package dev.idion.bladitodo.domain.dto;
 
-import dev.idion.bladitodo.domain.Board;
-import dev.idion.bladitodo.domain.List;
+import dev.idion.bladitodo.domain.entity.Board;
+import dev.idion.bladitodo.domain.entity.List;
 import java.util.stream.Collectors;
 import lombok.Builder;
 import lombok.Getter;
