+++ src/main/java/dev/idion/bladitodo/domain/dto/ListDTO.java
@@ -1,6 +1,6 @@
-package dev.idion.bladitodo.domain.http;
+package dev.idion.bladitodo.domain.dto;
 
-import dev.idion.bladitodo.domain.List;
+import dev.idion.bladitodo.domain.entity.List;
 import java.time.LocalDateTime;
 import java.util.stream.Collectors;
 import lombok.Builder;
